+++ week-7/assignment-1/.gitignore
@@ -1,3 +1,3 @@
-
+.env
 .idea/
 node_modules
