+++ week-7/assignment-1/middleware/index.ts
@@ -1,0 +1,26 @@
+import jwt from 'jsonwebtoken';
+import { Response, Request, NextFunction } from 'express';
+export const SECRET = 'SECr3t';  // This should be in an environment variable in a real application
+
+export const authenticateJwt = (req: Request, res: Response, next: NextFunction) => {
+  const authHeader = req.headers.authorization;
+  if (authHeader) {
+    const token = authHeader.split(' ')[1];
+    jwt.verify(token, SECRET, (err, user) => {
+      if (err) {
+        return res.sendStatus(403);
+      }
+      if(!user){
+        return res.sendStatus(403);
+      }
+      if(typeof user === "string"){
+        return res.sendStatus(403)
+      }
+      req.headers["userId"] = user.id;
+      next();
+    });
+  } else {
+    res.sendStatus(401);
+  }
+};
+
