+++ week-7/assignment-1/routes/auth.ts
@@ -1,7 +1,7 @@
-const jwt = require("jsonwebtoken");
-const express = require('express');
-const { authenticateJwt, SECRET } = require("../middleware/");
-const { User } = require("../db");
+import jwt from "jsonwebtoken";
+import express from 'express';
+import { authenticateJwt, SECRET } from "../middleware/";
+import { User } from "../db";
 const router = express.Router();
 
   router.post('/signup', async (req, res) => {
@@ -29,7 +29,7 @@ const router = express.Router();
   });
 
     router.get('/me', authenticateJwt, async (req, res) => {
-      const user = await User.findOne({ _id: req.userId });
+      const user = await User.findOne({ _id: req.headers["userId"] });
       if (user) {
         res.json({ username: user.username });
       } else {
@@ -37,4 +37,4 @@ const router = express.Router();
       }
     });
 
-  module.exports = router
+ export default router
